Grammar, spelling, etc.
Grammar, spelling, etc. I'd love to help more. I was looking for the android app code to fix up some of the interface issues. If someone can point me to that I'd be happy to help.

I think Kardia/Kai has a bright future and I'm very interested in helping however I can. This is just to get started and introduce myself. I'd like to be involved as much as I can.

Cheers!